refactor config:* commands to separate file
